# backend/swagger.yaml
openapi: 3.0.3
info:
  title: DataCircles CRM API
  description: |
    **DataCircles CRM + Invoicing Platform API Documentation**

    A comprehensive multi-tenant CRM and invoicing system with seat-based subscription management, OAuth authentication, and advanced document conversion capabilities.

    ## Key Features
    - Multi-tenant architecture with complete data isolation
    - OAuth 2.0 authentication (Google, Facebook, GitHub, LinkedIn) + Phone OTP
    - Seat-based subscription billing with prorated seat additions
    - Document conversion engine (Invoice ↔ Proforma ↔ Quotation ↔ Delivery Challan)
    - GST tax calculation (IGST/CGST+SGST)
    - Role-based access control (Admin/Staff)
    - Plan-based feature restrictions
    - Bulk import with CSV mapping

    ## Authentication
    All protected endpoints require authentication via:
    - **Bearer Token**: `Authorization: Bearer <jwt_token>` (for OAuth users)
    - **Phone Token**: `x-phone-token: <jwt_token>` (for phone-authenticated users)

    ## Rate Limiting
    - Standard endpoints: 100 requests per 15 minutes per IP
    - OTP endpoints: Special rate limiting to prevent abuse

    ## Pagination
    Most list endpoints support pagination with query parameters:
    - `page`: Page number (default: 1)
    - `limit`: Items per page (default: 20, max: 100)
    - `sort`: Sort field (prefix with `-` for descending)

    ## Multi-Tenancy
...
  version: 1.0.0
  contact:
    name: DataCircles Support
...
  license:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from OAuth or phone authentication
    PhoneToken:
      type: apiKey
      in: header
      name: x-phone-token
      description: JWT token for phone-authenticated users
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Items per page
...
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource ID
  responses:
    UnauthorizedError:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
paths:
...
  /super-admin/login:
    post:
      tags:
        - Super Admin
      summary: Super admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  superAdmin:
                    type: object
        "401":
          description: Invalid credentials
  /super-admin/dashboard:
    get:
      tags:
        - Super Admin
      summary: Super admin dashboard
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Dashboard data
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /super-admin/overview:
    get:
      tags:
        - Super Admin
      summary: Get overview metrics
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Overview metrics
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /super-admin/tenants:
    get:
      tags:
        - Super Admin
      summary: Get tenants (organizations) with pagination
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: List of tenants
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /super-admin/users:
    get:
      tags:
        - Super Admin
      summary: Get users with pagination
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: List of users
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /super-admin/organizations-filter:
    get:
      tags:
        - Super Admin
      summary: Get organizations for filter
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of organizations
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /super-admin/billing:
    get:
      tags:
        - Super Admin
      summary: Get billing overview
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Billing data
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /super-admin/analytics:
    get:
      tags:
        - Super Admin
      summary: Get analytics
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Analytics data
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /super-admin/support:
    get:
      tags:
        - Super Admin
      summary: Get support tickets
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of support tickets
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /super-admin/tenants/{id}:
    get:
      tags:
        - Super Admin
      summary: Get tenant details
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Tenant details
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    delete:
      tags:
        - Super Admin
      summary: Delete tenant (organization) by ID
      description: Deletes a tenant (organization) and all related data. **Super Admin only.**
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Organization and all related data deleted successfully
        "400":
          description: Invalid organization ID
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: Organization not found
  /super-admin/tickets:
    get:
      tags:
        - Super Admin
      summary: Get tickets
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of tickets
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      tags:
        - Super Admin
  /super-admin/users/{id}:
    delete:
      tags:
        - Super Admin
      summary: Delete user by ID
      description: Deletes a user by ID. **Super Admin only.**
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: User deleted successfully
        "400":
          description: Invalid user ID
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: User not found
  /super-admin/tickets/{id}:
    get:
      tags:
        - Super Admin
      summary: Get ticket by ID
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Ticket details
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
    put:
      tags:
        - Super Admin
      summary: Update ticket
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Ticket updated
        "400":
          description: Validation error
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
    delete:
      tags:
        - Super Admin
      summary: Delete ticket
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Ticket deleted
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /super-admin/support/{id}/status:
    post:
      tags:
        - Super Admin
      summary: Update ticket status
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
      responses:
        "200":
          description: Status updated
        "400":
          description: Validation error
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /super-admin/organizations/{orgId}/payments:
    get:
      tags:
        - Super Admin
      summary: Get organization payments
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of payments
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...